requirements.txt
tensorflow==2.20.0
streamlit
numpy
Pillow
